Separated routes.go and app.go to add easy-to-follow documentation in main.go
