+++ routes.go
@@ -8,6 +8,24 @@ import (
 	"time"
 )
 
+var (
+	maxRouteTime     = time.Second
+	validEntryPoints = map[string]struct{}{
+		"signup": {}, "signup-submit": {}, "signup-fail": {},
+		"login": {}, "login-submit": {}, "login-fail": {},
+		"dashboard": {}, "settings": {}, "delete-user": {},
+	}
+)
+
+func newMatchaRouter() *http.ServeMux {
+	mux := http.NewServeMux()
+	// TODO(@CarlosACJ55): Make a clean transition from the switch case to ServeMux
+	// mux.Handle("/{$}", http.TimeoutHandler(http.HandlerFunc(loadPage), maxRouteTime, ""))
+	mux.Handle("/public/", http.StripPrefix("/public", http.FileServer(http.Dir("public"))))
+	mux.Handle("/", http.TimeoutHandler(http.HandlerFunc(routeWithTimeout), maxRouteTime, ""))
+	return mux
+}
+
 func route(w http.ResponseWriter, r *http.Request) {
 	path := strings.TrimLeft(r.URL.Path, "/")
 	log.Println("Routing {" + path + "}")
@@ -29,26 +47,16 @@ func route(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// TODO(@FaaizMemonPurdue): This is an example of how go routines should be used, but we still need server timeouts
+// TODO(@FaaizMemonPurdue): This is an example of how go routines should be used, but we still need API call timeouts
 func routeWithTimeout(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(r.Context(), maxRouteTime)
 	defer cancel()
 	select {
 	case <-ctx.Done():
 		log.Println("Routing took longer than", maxRouteTime)
-	//case <-time.After(maxRouteTime):
 	default:
 		start := time.Now()
 		route(w, r)
 		log.Println("Routing done after", time.Since(start))
 	}
-}
-
-func matchaMux() *http.ServeMux {
-	mux := http.NewServeMux()
-	// TODO(@CarlosACJ55): Make a clean transition from the switch case to ServeMux
-	//mux.Handle("/{$}", http.TimeoutHandler(http.HandlerFunc(loadPage), maxRouteTime, ""))
-	mux.Handle("/public/", http.StripPrefix("/public", http.FileServer(http.Dir("public"))))
-	mux.Handle("/", http.TimeoutHandler(http.HandlerFunc(routeWithTimeout), maxRouteTime, ""))
-	return mux
 }
